added varnish cache frontend
Varnish config:
- homepage is cached for 1d
- /skin/* (toolbar stuff) is cached for 1d
- /catalog/* is cached for 1d
- homepage favicons
- all other requests are forwarded to kiwix-serve

Dockerfile updates:
- installed curl to fix is_alive kiwix-serve test (was broken)
- installed latest varnish version from upstream deb
- apt-install command split into multiple lines for readability
- added varnish clear cache on library update
- changed Dockerfile command to varnish

Having varnish as CMD is cleaner as we are now obeying docker convention to have one
main process for the image and this process responds friendly to stop requests (unlike kiwix-serve)

Entrypoint updates
- made start.sh the entrypoint instead of command (makes more sense)
- moved kiwix-serve startup/restart/watch to independent script (restart-kiwix-serve.sh)
- using printf instead of multiple echo for readability of cron.daily creations
- using a cron job for checking kiwix-serve everyminute instead of sleep loop

Kiwix-serve checker updates
- run on startup via cron
- run every minute via cron
- executes start() if timesout on test URL
- start() calls stop() before starting kiwix-serve
-

Killing kiwix-serve before starting it up ensures the start is useless in the case
the process did not die but got stuck. Previous version did not.

How to tell is a resource was cached?
Check the response's header, for cached resources there are two headers:

* `X-Varnish` which contains one or two numbers. If there are two numbers, the resource was
retrieved from the cache. Those represent the request IDs of the current and the one which created the cache entry.
* `Age`: contains the age of the resource in the cache (seconds)

So:

* Two `X-Varnish` IDs and `Age` greater than 0: result from cache
* Single `X-Varnish` ID and `Age: 0`: result from backend

There's no way to be sure if the resource is in the to-cache list or not except by making two requests
and verifying that at least the second came from the cache.