Use our mirror to check size
We make a HEAD request on ZIM files to find the actual size of the files.
Reusing the in-library URL, we'd end-up on whatever mirror while doing this.
We thus could end-up on HTTPS mirror while this doesn't handle it properly.
Make more sense to use the local server anyway (even though we go up the network stack)