Include service worker info to ideascube catalog (#181)
In order to adapt Kiwix Hotspot behavior based on Service Worker dependency, we need
the catalog to have that information in the first place.

This adds a new `sw:y` property to items that rely on service worker (ie. have the `_sw:yes` tag)

This was done specifically as to not increase much the size of this already quite large
catalog and the fact that we want to retire it soon in favor of OPDS.

Use this oportunity to update PyYAML and use the C libyaml parser to boost performances.