Fixed #194: fixed certs renewal for monitoring and drive
Root cause was the fact that monitoring is using a custom vhost file.
In this case, it is required to include /etc/nginx/vhost.d/default
in the configuration as this file is on the shared volume between
nginx-proxy and acme-companion and is responsible for exposing the acme-challenge path.

Using this oportunity to update our base images to the latest compatible version