Removed monitoring.openzim.org from HTTPS list
For some reason, monitoring.openzim.org being in the reverse-proxy's letsencrypt list
makes the certificate renewal request for the reverse-proxy to fail (third party
python code raises an exception).
Given it's not necessary for this particular domain and the coming switch to k8s,
removing it is a interval fix.