New netdata image with dynamic IP config
Zimfarm doesn't rely on a static list of worker IPs anymore, which was used in this
image.

We are now building the stream.conf (config of allowed IP to stream data to our netdata)
using the IPs returned by zimfarm's workers endpoint (which was edited to include IPs)
and a new ENVIRON to supplement it.

Because workers comes and go, the stream configuration must be updated regularily
and netdata restarted should it change. Image this now exec crond with a 15mn task
to check and update IP and conditionnaly restart netdata.