Rationalizing requests
- using the same session everywhere to ensure a consistent retry mechanism
- stop testing the torrent URL before sending to qbittorrent.
This was done because qBittorrent always responds OK regardless of functional URL.
This lead to issues with many requests (DNS, connection) that may or may not be host-related.
Maybe MB can cope with the amount of data but given it doesnt answers to HEAD not changes
its behavior based on how much is consumed, this induces load on it for no reason.
- poll longer (up to 10s) awaiting for a torrent to show up in qbittorrent after it being added.
This is in relation to above and because 1s proved to be too short during tests