Excluding torrent clients from logs
When importing raw HTTP logs into matomo,
we are now excluding all lines for which the user agent match a list of predefined
torrent clients.
The code looks for “excluded string” inside the “User-Agent” one which is why
our list consists only of names.
The list has been assembled from client names and not actual User-Agent